- Command Pattern Added - Command Pattern Move and Undo Mechanic Added - UI and Move Values Event Added when Undo

+++ Assets/Scripts/ScriptableData/PlayerData.cs
@@ -10,9 +10,9 @@ public class PlayerData : ScriptableObject
     
 
     //Direction
-    public int UpMove;
-    public int DownMove;
-    public int LeftMove;
-    public int RightMove;
+    public float UpMove;
+    public float DownMove;
+    public float LeftMove;
+    public float RightMove;
 
 }
